+++ package.json
@@ -19,6 +19,7 @@
     "eslint-import-resolver-typescript": "^2.5.0",
     "expo": "~42.0.0",
     "expo-app-loading": "1.1.2",
+    "expo-camera": "^12.0.3",
     "expo-font": "~9.2.1",
     "expo-status-bar": "~1.0.4",
     "expo-web-browser": "~9.2.0",
