+++ src/screens/Forms/FormsScreen.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import GeneralQuestionManager from 'components/GeneralQuestionManager/GeneralQuestionManager';
 import UploadScreen from 'screens/Upload/UploadScreen';
 import { ScrollPageContainer, PageContainer } from 'screens/styles';
+import CameraScreen from 'screens/Upload/UploadScreen';
 
 const FormsScreen = () => {
   return (
