+++ src/screens/Upload/UploadScreen.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable react/style-prop-object */
-import { Camera } from 'expo-camera';
-import React, { useState } from 'react';
+import firebase from 'database/clientApp';
+import { Camera, CameraCapturedPicture } from 'expo-camera';
+import React, { useState, useEffect } from 'react';
 import {
   Button,
   ImageBackground,
@@ -8,6 +9,9 @@ import {
   TouchableOpacity,
   View,
 } from 'react-native';
+import { Document, Client } from 'types/types';
+import { setDocument } from 'database/queries';
+import { getCurrentClient } from 'database/auth';
 
 const styles = StyleSheet.create({
   container: {
@@ -36,66 +40,84 @@ const styles = StyleSheet.create({
     display: 'flex',
     backgroundColor: '#fff',
   },
+  previewContainer: {
+    flex: 1,
+    width: '100%',
+  },
   previewImage: {
     width: '100%',
     height: '100%',
-    padding: 100,
   },
 });
 
-// states to act as pseudo-navigators
 enum CameraStatus {
   INACTIVE, // inactive camera
   ACTIVE, // open camera
   PREVIEW, // viewing taken picture
 }
 
-// photo type
-type Photo = {
-  uri: string;
-  height: number;
-  width: number;
-};
-
 export default function UploadScreen() {
   let camera: Camera | null;
+  const storage = firebase.storage();
   const [cameraState, setCameraState] = useState<CameraStatus>(
     CameraStatus.INACTIVE,
   );
-  const [photo, setPhoto] = useState<Photo | undefined>(undefined);
+  const [photo, setPhoto] = useState<CameraCapturedPicture | undefined>(
+    undefined,
+  );
+  const [client, setClient] = useState<Client | undefined>(undefined);
 
-  // opens camera & lets user take photo
+  const loadClient = async (): Promise<void> => {
+    const c: Client | undefined = await getCurrentClient();
+    setClient(c);
+  };
+
+  useEffect(() => {
+    loadClient();
+  }, []);
+
   const startCamera = async () => {
     const { status } = await Camera.requestCameraPermissionsAsync();
     if (status === 'granted') {
-      // navigates to camera screen
       setCameraState(CameraStatus.ACTIVE);
     } else {
       // TODO: access denied -> go back?
     }
   };
 
-  // takes picture through camera, sets photo as Photo state
   const takePicture = async () => {
     if (!camera) {
       return;
     }
-    const currentPhoto: Photo = await camera.takePictureAsync();
+    const currentPhoto: CameraCapturedPicture = await camera.takePictureAsync({
+      base64: true,
+    });
     setPhoto(currentPhoto);
-    setCameraState(CameraStatus.PREVIEW);
   };
 
-  // renders camera preview (displaying photo prop)
-  const CameraPreview = ({ photo2 }: any) => (
-    <View>
-      <ImageBackground
-        source={{ uri: photo2 && photo2.uri }}
-        style={styles.previewImage}
-      />
-    </View>
-  );
+  useEffect(() => {
+    if (photo !== undefined) {
+      setCameraState(CameraStatus.PREVIEW);
+    }
+  }, [photo]);
 
-  // pseudo navigation (renders different things based on state of CameraStatus)
+  const uploadPicture = async () => {
+    const reference = storage.ref();
+    const ref2 = reference.child(`${client!.id}.jpg`);
+    await ref2.putString(photo?.uri!, 'data_url', {
+      contentType: 'image/jpg',
+    });
+    const url = await ref2.getDownloadURL();
+    // TODO after Noah finishes document sprint
+    // const thisPhoto: Document = {
+    //   id: 'sample-pic', // set this to some id
+    //   url,
+    //   type: 'passport', // check what this should be
+    //   createdAt: new Date(),
+    // };
+    // setDocument('clientId', 'caseId', thisPhoto);
+  };
+
   switch (cameraState) {
     case CameraStatus.ACTIVE:
       return (
@@ -115,7 +137,15 @@ export default function UploadScreen() {
         </View>
       );
     case CameraStatus.PREVIEW:
-      return <CameraPreview photo={photo} />;
+      return (
+        <View style={styles.previewContainer}>
+          <ImageBackground
+            source={{ uri: photo!.uri }}
+            style={styles.previewImage}
+          />
+          <Button title="Keep Scan" onPress={uploadPicture} />
+        </View>
+      );
     default:
       return (
         <View style={styles.container}>
